// src/features/ai/ops/types.ts
export interface DocDiff {
  added: string[];
  changed: string[];
  removed: string[];
}